move PhotonAttnProcessor2_0 in transformer_photon
